fixup 2: take care of similarly named fns in different schemas
In the (possibly unlikely) case where in the future we introduce a
function under `crdb_internal.` with the same name as another
function under `pg_catalog.`, we need to distinguish them.

Also, the result of the completion generation should now
properly quote the identifiers, since we're allowing this to
complete UDFs now and UDFs can have fancy names.